+++ src/__tests__/Store-test.js
@@ -11,22 +11,6 @@ describe('Store', () => {
 
   let actionId = 'actionId';
 
-  describe('#getState()', () => {
-
-    let s = new ExampleStore();
-
-    it('returns state object', () => {
-      expect(s.getState()).to.deep.equal({ foo: 'bar' });
-    });
-
-    it('prevents mutations of state object', () => {
-      let state = s.getState();
-      state.foo = 'changed';
-
-      expect(s.getState()).to.deep.equal({ foo: 'bar' });
-    });
-  });
-
   describe('#register()', () => {
     it('adds handler to internal collection of handlers', () => {
       let store = new ExampleStore();
@@ -75,6 +59,11 @@ describe('Store', () => {
       expect(store.register.bind(store, null)).not.to.throw();
     });
 
+  });
+
+  it('default state is null', () => {
+    const store = new Store();
+    expect(store.state).to.be.null;
   });
 
   describe('#registerAsync()', () => {
@@ -191,6 +180,42 @@ describe('Store', () => {
     });
   });
 
+  describe('#forceUpdate()', () => {
+    it('emits change event', () => {
+      let store = new ExampleStore();
+      let listener = sinon.spy();
+      store.addListener('change', listener);
+
+      store.forceUpdate();
+
+      expect(listener.calledOnce).to.be.true;
+    });
+
+    it('doesn\'t modify existing state', () => {
+      let store = new ExampleStore();
+      let listener = sinon.spy();
+      store.addListener('change', listener);
+
+      store.register(actionId, function() {
+        this.replaceState({ bar: 'baz' });
+        this.forceUpdate();
+
+        expect(this.state).to.deep.equal({ foo: 'bar' });
+        expect(listener.called).to.be.false;
+
+        this.setState({ foo: 'bar' });
+        this.forceUpdate();
+        this.replaceState({ baz: 'foo' });
+      });
+
+      // Simulate dispatch
+      store.handler({ actionId, body: 'foobar' });
+
+      expect(listener.calledOnce).to.be.true;
+      expect(store.state).to.deep.equal({ baz: 'foo' });
+    });
+  });
+
   describe('#setState()', () => {
     it('shallow merges old state with new state', () => {
       let store = new ExampleStore();
@@ -201,6 +226,17 @@ describe('Store', () => {
         foo: 'bar',
         bar: 'baz',
       });
+    });
+
+    it('supports transactional updates', () => {
+      const store = new Store();
+      store.state = { a: 1 };
+      store.setState(state => ({ a: state.a + 1 }));
+      expect(store.state.a).to.equal(2);
+      store.setState(state => ({ a: state.a + 1 }));
+      expect(store.state.a).to.equal(3);
+      store.setState(state => ({ a: state.a + 1 }));
+      expect(store.state.a).to.equal(4);
     });
 
     it('emits change event', () => {
@@ -290,6 +326,30 @@ describe('Store', () => {
       store.replaceState({ foo: 'bar' });
 
       expect(listener.calledOnce).to.be.true;
+    });
+  });
+
+  describe('.assignState', () => {
+    it('can be overridden to enable custom state types', () => {
+      class StringStore extends Store {
+        static assignState(prevState, nextState) {
+          return [prevState, nextState]
+            .filter(state => typeof state === 'string')
+            .join('');
+        }
+      }
+
+      const store = new StringStore();
+
+      expect(store.state).to.be.null;
+      store.setState('a');
+      expect(store.state).to.equal('a');
+      store.setState('b');
+      expect(store.state).to.equal('ab');
+      store.replaceState('xyz');
+      expect(store.state).to.equal('xyz');
+      store.setState('zyx');
+      expect(store.state).to.equal('xyzzyx');
     });
   });